Make the direct-composition demo compile to an empty program on non-windows.
