+++ direct-composition/Cargo.toml
@@ -4,11 +4,9 @@ version = "0.1.0"
 authors = ["Simon Sapin <simon.sapin@exyr.org>"]
 
 [target.'cfg(windows)'.dependencies]
-winapi = {version = "0.3", features = ["winerror", "d3d11", "dcomp"]}
-
-[dependencies]
 euclid = "0.17"
 gleam = "0.4"
 mozangle = {version = "0.1", features = ["egl"]}
 webrender = {path = "../webrender"}
+winapi = {version = "0.3", features = ["winerror", "d3d11", "dcomp"]}
 winit = "0.10"
